requirements.txt
Flask
gunicorn
instagram_web_api
sightengine
flask
future
python-twitter
PySocks
tweepy
